refactor: 정보 작성페이지 파일 이동
