+++ src/components/SelectEventInfo/SelectEventInfo.tsx
